@@ -1,5 +1,5 @@
 {
-  "name": "vue-codemod",
+  "name": "@cldcvr/vue-codemod",
   "version": "0.0.5",
   "description": "Vue codemod scripts",
   "main": "dist/index.js",
@@ -21,32 +21,38 @@
   "author": "Haoqun Jiang",
   "license": "MIT",
   "dependencies": {
-    "@babel/core": "^7.10.3",
-    "@babel/preset-env": "^7.10.3",
-    "@babel/types": "^7.12.12",
-    "@types/jscodeshift": "^0.7.1",
-    "@vue/compiler-core": "^3.0.5",
-    "@vue/compiler-dom": "^3.0.5",
+    "@babel/core": "^7.18.6",
+    "@babel/preset-env": "^7.18.6",
+    "@babel/types": "^7.18.7",
+    "@types/jscodeshift": "^0.11.5",
+    "@vue/compiler-core": "^3.2.37",
+    "@vue/compiler-dom": "^3.2.37",
     "debug": "^4.1.1",
     "globby": "^11.0.2",
-    "jscodeshift": "^0.11.0",
+    "jscodeshift": "^0.13.1",
     "lru-cache": "^6.0.0",
-    "source-map": "^0.6.1",
+    "source-map": "^0.7.4",
     "yargs": "^16.2.0"
   },
   "devDependencies": {
+    "@types/babel__core": "^7.1.19",
+    "@types/babel__preset-env": "^7.9.2",
+    "@types/concurrently": "^6.4.0",
     "@types/debug": "^4.1.5",
-    "@types/jest": "^26.0.3",
+    "@types/jest": "^28.1.4",
+    "@types/koa": "^2.13.4",
+    "@types/lru-cache": "^7.10.10",
     "@types/node": "^12.12.47",
+    "@types/prettier": "^2.6.3",
     "@types/yargs": "^15.0.4",
     "@vitejs/plugin-vue": "^1.6.2",
     "concurrently": "^6.2.1",
-    "jest": "^26.1.0",
+    "jest": "^28.1.2",
     "koa": "^2.13.1",
-    "prettier": "^2.0.4",
-    "ts-jest": "^26.1.1",
-    "ts-node-dev": "^1.1.8",
-    "typescript": "^4.1.3",
+    "prettier": "^2.7.1",
+    "ts-jest": "^28.0.5",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^4.7.4",
     "vite": "^2.5.6"
   },
   "jest": {
@@ -55,11 +61,13 @@
     "testPathIgnorePatterns": [
       "<rootDir>/dist/",
       "<rootDir>/node_modules/"
-    ]
-  },
-  "prettier": {
-    "semi": false,
-    "singleQuote": true
+    ],
+    "globals": {
+      "ts-jest": {
+        "diagnostics": false,
+        "isolatedModules": true
+      }
+    }
   },
   "engines": {
     "node": ">= 10.0"
